requirements.txt
flask==2.2.3
requests==2.28.2
beautifulsoup4==4.11.2
nltk==3.8.1
scikit-learn==1.2.2
pandas==1.5.3
newspaper3k==0.2.8
pyttsx3==2.90
gTTS==2.3.1
python-dotenv==1.0.0 